Upload_Github/requirements.txt
pandas==2.1.4
numpy==1.26.4

sentence-transformers==2.6.1
torch==2.2.2

faiss-cpu==1.7.4

google-generativeai==0.5.4

tqdm==4.66.2
scikit-learn==1.4.0